+++ R/zzz.R
@@ -1,7 +1,8 @@
 .onLoad <- function(libname, pkgname) {
   opts <- options()
   opts.quickPlot <- list( # nolint
-    quickPlot.tolerance = .Machine$double.eps ^ 0.5
+    quickPlot.tolerance = .Machine$double.eps ^ 0.5,
+    quickPlot.maxNumPolygons = 3e4
   )
   toset <- !(names(opts.quickPlot) %in% names(opts))
   if (any(toset)) options(opts.quickPlot[toset])