+++ NEWS.md
@@ -2,7 +2,13 @@ Known issues: https://github.com/PredictiveEcology/quickPlot/issues
 
 version 0.1.4
 =============
+* `clickExtent` now more interactive with new examples
+* `SpatialPolygons` get a fill colour by default if not passed in by user
 * fix bug related to plotting with NAs (#14, @CeresBarros)
+* much faster for `SpatialPolygons`
+* bug fixes
+* new options("quickPlot.maxNumPolygons"), used for plotting of large `SpatialPolygons` objects
+ 
 
 version 0.1.3
 =============